Renamed buttonWithIcon.tsx to button-with-icon.tsx
